+++ validators/topology_snapshot_test.go
@@ -37,6 +37,22 @@ func TestEmptySnapshot(t *testing.T) {
 	assert.Equal(t, 1, len(top.Keys()))
 }
 
+func TestChangeOnValidatorPerfUpdate(t *testing.T) {
+	top := getTestTopology(t)
+	defer top.ctrl.Finish()
+
+	h, err := top.GetHash(topKey)
+	assert.Nil(t, err)
+	assert.NotEmpty(t, h)
+
+	updateValidatorPerformanceToNonDefaultState(t, top.Topology)
+
+	h2, err := top.GetHash(topKey)
+	assert.Nil(t, err)
+	assert.NotEmpty(t, h2)
+	require.NotEqual(t, h, h2)
+}
+
 func TestTopologySnapshot(t *testing.T) {
 	top := getTestTopWithDefaultValidator(t)
 	updateValidatorPerformanceToNonDefaultState(t, top.Topology)
@@ -45,7 +61,7 @@ func TestTopologySnapshot(t *testing.T) {
 	h1, err := top.GetHash(topKey)
 	require.Nil(t, err)
 
-	tmPubKeys := []string{"tm-pubkey-1", "tm-pubkey-2"}
+	tmPubKeys := []string{"2w5hxsVqWFTV6/f0swyNVqOhY1vWI42MrfO0xkUqsiA=", "67g7+123M0kfMR35U7LLq09eEU1dVr6jHBEgEtPzkrs="}
 	ctx := context.Background()
 
 	nr1 := commandspb.AnnounceNode{